fixing nav bug with getStaticProps and setting up global context"
